pyproject.toml
[project]
name = "bot"
description = "a bluesky bot"
readme = "README.md"
authors = [{ name = "zzstoatzz", email = "thrast36@gmail.com" }]
requires-python = ">=3.12"
dynamic = ["version"]
dependencies = [
    "anthropic",
    "atproto",
    "fastapi",
    "httpx",
    "openai",
    "pydantic-ai",
    "pydantic-settings",
    "rich",
    "turbopuffer",
    "uvicorn",
]

[tool.hatch.version]
source = "vcs"

[tool.ruff.lint]
extend-select = ["I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.uv]
dev-dependencies = [
    "pytest-sugar",
 "pytest-asyncio",
 "ruff",
 "ty",
]


[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"
